componentize for legibility and increase perf
